move `absolute_path`, and `relative_path` to separate module `io.file_utils`
