+++ src/dfastbe/gui/gui.py
@@ -32,7 +32,8 @@ from PyQt5 import QtWidgets
 from PyQt5 import QtCore
 import PyQt5.QtGui
 
-from dfastbe.io.io import get_text, absolute_path, ConfigFile
+from dfastbe.io.file_utils import absolute_path
+from dfastbe.io.io import get_text, ConfigFile
 import pathlib
 from pathlib import Path
 import sys
